move function to plot utils
